+++ src/EditMultipleSoundProducersDialog.h
@@ -44,9 +44,9 @@ private:
 	
 	//sound producer attributes
 	std::string name;
-	double xPosition;
-	double yPosition;
-	double zPosition;
+	float xPosition;
+	float yPosition;
+	float zPosition;
 	bool tempFreeRoamBool;
 	std::uint8_t sound_bank_account_num;
